FunctionExecutor: provide a switching mechanism for WasmEdge and wasmtime
