+++ function-executor/Cargo.toml
@@ -9,9 +9,11 @@ license = "Apache-2.0"
 [dependencies]
 anyhow = "1.0.45"
 eyre = "0.6.5"
-wasmedge-sdk = "0.3.0"
-wasmedge-sys = "0.8.0"
-wasmedge-types = "0.2.0"
+wasmtime = { version = "0.33.0", optional = true }
+wasmtime-wasi = { version = "0.33.0", optional = true }
+wasmedge-sdk = { version = "0.3.0", optional = true }
+wasmedge-sys = { version = "0.8.0", optional = true }
+wasmedge-types = { version = "0.2.0", optional = true }
 argh = "0.1.6"
 # keep in sync with anna-rs
 zenoh = { git = "https://github.com/eclipse-zenoh/zenoh.git", rev = "d4b00540cd0faa6ce585e11862cf9740ca226489" }
@@ -24,3 +26,8 @@ chrono = "0.4.19"
 serde = { version = "1.0.130", features = ["derive"] }
 bincode = "1.3.3"
 anna = { git = "https://github.com/essa-project/anna-rs.git" }
+
+[features]
+default = ["wasmedge_executor"]
+wasmtime_executor = ["wasmtime", "wasmtime-wasi"]
+wasmedge_executor = ["wasmedge-sdk", "wasmedge-sys", "wasmedge-types"]
